requirements.txt
Flask==3.0.0
flask-sock==0.7.0
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper>=20231117
requests>=2.28.0
soundfile>=0.12.0